requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
APScheduler==3.10.4
srt==3.5.3
numpy>=2.0.2,<3.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
google-cloud-storage==2.14.0
psutil==5.9.8
boto3==1.34.34
Pillow>=10.0.0
matplotlib>=3.9.0
yt-dlp==2023.12.30
jsonschema>=4.17.3
ffmpeg-python>=0.2.0
backoff>=2.2.1
youtube-transcript-api>=1.0.3
redis>=4.5.0
rq>=1.16.1
click>=8.1.7
emoji>=2.0.0
openai-whisper>=20240930
playwright>=1.40.0
pillow>=10.3.0
edge-tts>=6.1.0
